+++ sdk/storage/Azure.Storage.DataMovement/src/TransferManager.cs
@@ -132,13 +132,13 @@ namespace Azure.Storage.DataMovement
         /// Return true once the transfer has been successfully paused or false if the transfer
         /// was already completed.
         /// </returns>
-        public virtual async Task PauseTransferIfRunningAsync(string transferId, CancellationToken cancellationToken = default)
+        public virtual async Task PauseTransferAsync(string transferId, CancellationToken cancellationToken = default)
         {
             cancellationToken = LinkCancellation(cancellationToken);
             Argument.AssertNotNullOrEmpty(transferId, nameof(transferId));
             if (!_transfers.TryGetValue(transferId, out TransferOperation transfer))
             {
-                throw Errors.InvalidTransferId(nameof(PauseTransferIfRunningAsync), transferId);
+                throw Errors.InvalidTransferId(nameof(PauseTransferAsync), transferId);
             }
             await transfer.PauseAsync(cancellationToken: cancellationToken).ConfigureAwait(false);
         }